tools: integrate the if_rate project for live network monitoring
if_rate is a very simple utility capable of returning both the bit rate
and the packet rate in each direction of an arbitrary list of network
devices, at a specified interval. The output format is suitable for
graphing and contains the absolute date to ease synchronization with
other metrics sources.