scripts: add missing set-irq.sh
This is a tiny but convenient script that looks for a NIC name (or any
word) in /proc/interrupts, figures all associated interrupts and spreads
them over a number of cores starting from the top. The main use case is
to bind network IRQs on the upper CPUs and leave the lower ones available
for processes.

No consideration is taken of SMT nor nodes. A mask ought to be added to
restrict the usage to ta subset of CPUs but for now it remains simple.